feat: split into getSwcLoaderOptions and getJsTransformRules
