+++ packages/repack/src/utils/index.ts
@@ -4,4 +4,5 @@ export * from './getDirname.js';
 export * from './getPublicPath.js';
 export * from './getResolveOptions.js';
 export * from './getModulePaths.js';
-export * from './getJSTransformRules.js';
+export * from './getJsTransformRules.js';
+export * from './getSwcLoaderOptions.js';
